refactor(ci): reduce branch pushes to refreshing the dev environment
A push to a default or release branch replayed the whole battery: the six per-module suites,
the five frontend lint-and-test jobs, the e2e suites, the chainguard and chainguard-fips images,
and the community build. Fifty-eight jobs to refresh an environment. A pull request has already
run what its changes could affect — the changed-files predicates are dependency-aware — so the
push now builds, publishes and deploys, and nothing else. Seventeen jobs, sixteen on a release
branch, which has no NextGen deployment.

Three dependency lists had to be rewired, all pointing at jobs that no longer run on this path:

  - Publish on artifactory and Publish on nexus waited on the six per-module test jobs; they
    wait on the build instead.
  - Deploy on Azure cluster waited on those six plus the four product images; it waits on the
    images.
  - The Aikido scans were generated asking for the chainguard and chainguard-fips images, so
    they would have waited forever on jobs absent from this workflow. They now cover the
    variants this path actually pushes.

Left behind them: dev-environment-jobs, the shared build set the scheduled workflow also
composes — verified to leave it with the same sixteen jobs and the same dependencies — and
publish-and-deploy-jobs, which is master-and-support-jobs renamed to what it now does.
chainguard-fips-jobs is deleted: its only caller was this path, and the manual workflow builds
its own versioned images.

Every requires in all fifty-one fixtures was checked to resolve to a job present in its
workflow.
